requirements.txt
fastapi
uvicorn
pypdf
sentence-transformers
python-dotenv
requests